- Encrypte and Decrepte functions now have there own classes. - logger class added.
